change: 物体検出によって出力されたすべてのroiを表示するように変更
